SeniorCare-main/seniorcare_backend/requirements.txt
flask
firebase_admin
gunicorn
python-dotenv
python-dateutil
